Dont allow beam when training is on
This change prevents a beam from being added when
we are training.  This makes it easier to share a config
between the training and exporter, where the latter will
likely need to specify a beam width.  The exporter already
automatically adds in a predict=True to the configuration.
If we didnt do this, we would have to support this as a CL
option to export, which is not idea, since we want to share
the same script between tasks