move the vocab outside of the models, add services
The services access the 3 components: 1. vocab 2. vectorizers
and 3. model.  They can use these to easily perform inference
without the hassle required in previous versions.

The model is now separate from the other 2 components, and
only knows vectors.  The api-examples are updated to show how
inference now works.

The vectorizers currently use pickle as a save and load format,
the vocabs use JSON (as before), and the model is as before, but
w/o the vocabs.  The mead harness writes each component as soon
as its feasible to capture them, and ultimately zips them using
the `basedir`.

Also updated `stacked_cnn` to be reusable from relevant scoping