fix sharing issues for multi-GPU+inference model
we were sharing a single embeddings object between all
multi-GPU graphs.  This was ok (though perhaps unclean)
in the training code, but in seq2seq this was causing
the embedding.x input to be shared between all models.
This change adds a `detached_ref()` method which is
basically a copy-constructor so that .x is no longer bound