change call that required pd>=0.21.0
Previously, `xpctl` called the `drop(columns=X`)
method.  This required the pandas version to be
0.21 or greater.  Since this is a one-liner and
can be fixed easily, lets be lenient on the pd
requirement and just change that line.

Also make the config logging more explicit.
if someone fat-fingers the CL, they may currently
not realize that the file they requested wasnt on
the FS.  The logging here identifies the path that
is taken in an attempt to help the user realize
that they probably didnt intent to provie a config
string.