Allow overriding the dataset index with user files
This change allows overriding the dataset index by
providing a separate CL options for modified
training set, validation set and test set.

If the mead config has a matching label, that is used
as a template and the CL args are treated as an
overlay yielding a new dataset entry.  If no label
exists, a new entry is created in the index
for testing and debugging purposes.

The purpose of this option is to allow a user
to do simple things like debug by intentionally
overfit a file using an existing mead config
as a template without having to add those testing
datasets into the index at all