feature: lm src keys (#302)
* use `src_keys` to decide features to use in LM

Our current code has no way to specify only e.g. char
features when we predict words.  This is because we
require words for target keys and we dont have any
conditional logic to avoid the target key in embeddings.

IOW, its currently possible to have a word LM, a char LM,
a word+char -> word/char LM, but its not possible to have
a char -> word LM (like the Kim model).  This change adds
a `src_keys` field, which defaults to all features defined,
that allows a user to mask which keys should be used as input
features

* oops

* fix pytorch version