TF by default steals all GPU memory for a device
This is causing issues on an RTX 2080 with TF 1.13
and cuDNN 7.6 (with CUDA 10).  The issue is documented
here: https://github.com/tensorflow/tensorflow/issues/24496
and the only fix at this time appears to be to tell TF in
software not to do this.

I created a wrapper function called `create_session()`
which does this, and changed all instances of `tf.Session()`
(with no args) to use that