this change lazily initializes the pytorch optz
only in the case where the optimizer isnt previosly
injected.  this gets around the issue of the model parameters
being needed first.  The caller can either formally initialize
the optimizer in the usual pytorch way and inject it into
the EagerOptimizer, or pass in the named parameters, in which
case it will defer until the first update is called