TF 2.0 works in eager mode (#361)
* eager.py runs

* almost mapped (Bi)LSTM interfaces between 1 and 2

* try to account for difference of interface

* eager.py runs

* almost mapped (Bi)LSTM interfaces between 1 and 2

* clean up the TF set a bit

* update APIs

* update pytorch example to use datasets

* rename, make more similar

* more similar

* OptimizerManager -> torch.optim.Adam

* redundant now

* move stuff around so its similar

* eager.py runs

* normalize layers between tf and pytorch

this means a lot of TF layers end up with a new
input size first argument, which can be passed None
if desired

* rebase again

* move each training hook into its own sub-module

conditionally import the riht stuff based on TF version
when we are in TF2.0 we have to do a bunch of changes
to the model, particularly regarding placeholder removal
model loading and saving for 2.0 is TBD here.

* make tagger training run in eager mode with 2.0

the issue about momentum not working is problematic
since most of the configs rely on this setting

* this change lazily initializes the pytorch optz

only in the case where the optimizer isnt previosly
injected.  this gets around the issue of the model parameters
being needed first.  The caller can either formally initialize
the optimizer in the usual pytorch way and inject it into
the EagerOptimizer, or pass in the named parameters, in which
case it will defer until the first update is called

* fix eager LM to run again

* oops, fix straggler

* fix: grads and variables could get unaligned (#368)

* Add LSTM encoder with state (tf2)

* initial add of the notes for v2

* turn off required lengths

* use the SET_TRAIN_FLAG()

* fix: get the other tf classify models running (#369)

* Added in eager LM

* add additional function hook for LMs

* chore: add back the length aware mean pooling (#370)