chore stop illegal outputs in the last position of the tf tagger
In IOBES tagging there are only a subset of tags that can be the
last tag in a sequence, for example you can't end on a B or I tag
because it will result in a span that doesn't end with an E. In
our constrained decoding we handle this by having a set of illegal
transition from tags to the EOS symbol. In our pytorch CRF the
transition to EOS is calculated inside. In tf we use the built in
CRF which doesn't include transitions to EOS inside the CRF. We
fix that here by inserting EOS tokens explicitly into the emission
scores.