fix: labels dont normally have pad (#460)
I think this pop was introduced for the bpe tagger where a special label vectorizer will output pad labels for the tokens that are not the start of a bpe word. The problem is that for normal sequence tagging tasks the pad symbol isnt normally in the label space so this pop causes and error. This should fix it