fix: use an output delimiter when the input is None (#466)
* fix: use an output delimiter when the input is None
When you read a conll file you can use None to properly parse both tab delimited and space delimited file but when
the delimiter is None the writing fails. This code adds a guard so that if you have a None delimiter coming in it
will use a space as the output delimiter. If you pass in a delimiter it will use that as the output one

* chore: test that would have caught the None bug