fix: use if not elif (#473)
In the _optimize_loss function we have an if statement that checks if
there is a learning rate decay function. If there is we initialize it
with the learning rate and then call that function the lr so that if we
pass it into an optimizer it will be used which training. The problem
was that the first if statement in the block of `elif`s that dealt with
creating the optimizer object. Being an `elif` it wouldn't execute if we
hit the block about the learning rate decay so the `opt` would never
actually be created. I fixed this by switching it to an `if`. The two
sets of conditionals are doing different things so they shouldn't have
been one large block of `if` and `elif`s