fix bug in tied s2s where dst wasnt set right
the code was trying to use n-1 tokens which causes
the encoder and decoder tensor lengths to be different.
this is an error as the transformer code, particularly
where relative attention is used expects them to be padded
to the same length (this is consistent with T2T).

the DST buffer should be padded with a <GO> token on the
front, and the TGT buffer should be the sequence to predict.