feat: Added tools and models for calibration (#484)
* feat: Added tools and models for calibration

Calibration is an important tool for trustworthy models. When you have a
classifier it often outputs scores for each class that people
interpret as probabilities for each class. The problem is these are
often not well calibrated. This means that when you have 100 examples
that all have a score of 0.8 we would expected that about 80 of those
100 examples would be correctly labeled by the model, however; with
neural networks this is often not the case. When a model says that
examples have a probability of 0.8 the proportion of examples that are
correct will be much less than than the expected 80/100. These models
are over confident in their answers.

This PR adds metrics, models, and training for calibration as discussed
in [On Calibration of Modern Neural Networks](https://arxiv.org/abs/1706.04599).
It adds metrics and visualization as an `eight_mile.calibration`
package, adds various post-hoc methods for calibration as addons, adds
and api-example to collect metrics on calibrations, and adds a script
that can combine the various graphs that describe calibration into a
single figure.

The way this models are trained is with a checkpoint. First you train a
model as normal with `mead-train` and then you use a second config that
specifies the same model but set the model type to one of the new
calibrations methods and pass the checkpoint of the old model in. You
need to make sure the model is specified the same as the old model or
else weights won't get read into the checkpoint.

The final thing this PR does it that it updates the classify trainer to
only update the `.trainable_variables` of a model. This means that by
controlling this value in the post-hoc calibration models we can in
effect freeze the rest of the network weights by ignoring them in the
optimization.

* chore: updated based on comments

* chore: remove _get_batchsz