Fix/tf1 eager lr sched (#500)
* add an optional argument for num_replicas

* fix LR sched to be cond on eager, not TF version

previously, the LR scheduler API was being conditioned
on whether TF was version 1.x or 2.x.  This is incorrect,
the signature for the LR scheduler changes depending on
whether we are using the Keras API or TF API. Currently,
we only use the Keras API when we are in eager mode, no
matter if we are using TF 1.x or 2.x.

This change sets the LR scheduler API conditioned on
the TF eager mode instead of version.