chore: function that remove common extensions for conll files (#516)
I have written a similar function that can strip off the conll
extensions like .conll or .iobes multiple times across various projects.
This is adding that function to baseline so we can reuse it instead of
copying it around everywhere. I made the decision to make the extension
removal case sensitive so like linux paths names are, we might want to
change that in the future.