add Reduction component that EmbeddingsStack uses
the reduction operator allows the stack to reduce to
a sum, which makes it possible to do things like
composing multiple embedding types like positional,
token and token types as separate features.

For cases where there are multiple embeddings that
would normally be joined with concatenation, this also
makes it possible to do other reductions.

Currently, the change here doenst use a registry
which would make it easier to extend the layer
reduction but this could be added later