for BERT checkpt load, transform TF needs LN after
BERT is based on the original T2T architecture,
where the layer norms happen after the transformation.

This makes that optional, and also allows setting eps
in the layer norm to make sure that BERT checkpoints
can be loaded.

Also RA attention needs to be supported in decoders,
for some reason that never got added.