Remove 8mi models as internal calls from baseline (#542)
* Remove 8mi models as internal calls from baseline

make same classify interface between PyTorch/TF

also I updated the comments and added documentation
on the MEAD-Baseline classify architecture to
docs/classify.md

normalize and unify the implementations TF/PyTorch

Tagger ABC follows same flow as Classifier ABCs

update docs on vectorizers

more docs

* remove copy-paste code

* fix WS

* remove old function

* add init_proj function to TransducerTaggerModel

this adds an init function for the projection layer.
if you dont need it or if you want to use a res. conn
because the output of the encoder is the same size, you
might want to override this.  OW, leave it alone

* fix type hints

* DRY, add init_output hook to fine-tune models

* updated

* update fine-tuning example with mead-hub

* TransducerTaggerModel-> AbstractEncoderTaggerModel