chore: default batchsz to be in train section (#552)
This PR updates the various slightly divergent implementations of
`._reorganize_params()` Some allowed the batch size to be in the train
section some didn't. Some were looking in the main params for the test
batch size, some weren't.

This PR makes it so that a `batchsz` key in the train section will
be used first with a back off to the `batchsz` key in the whole params.
If no `batchsz` is specified at all then a KeyError is thrown.

This PR makes it so that a `test_batchsz` in the train section will be
used first with a back off to a `test_batchsz` key in the main parameter
section. If neither of these are defined a size of `1` is used.