add a MEAD_HUB_MODULES global and persist to state
this modification adds a MEAD_HUB_MODULES variable into
baseline.utils and whenever a module is downloaded, it adds
the module.  When the state is recorded for embeddings or for
models, this is persisted, and when the model rehydrates,
it first downloads any hub modules that are persisted and not stored.

In PyTorch, this is not currently required as the full model is
persisted to the pytorch serialized state