Fix/eager es option (#580)
* using eval set supposed to be optional.  fix eager

previously, the eager mode `fit_func`s did not have
an option that allowed them not to run the test set
even though it can be None.  A simple fix is to put
the dataset loading in the portion for testing

* the datasets fit_funcs also had the problem

additionally, we have come to rely on having an
output_dim but we were missing some. fixed those