support non-directory/non-zip checkpoint names
sometimes there might be multiple pid names in
the same file. this happens when multiple runs are
done and then the user needs to load only one.

by checking the job status or even just the time
of each checkpoint written out, the user can easily
determine the PID-specific job name, but its not
currently loadable.

This change allows the user to give a longer expression
as the checkpoint name and resolves that to the full
checkpoint name, and passes this along to the suffix