feat: a utility to write out a conll file from a list of sentences
This isn't as extensive as the utilties for reading in conll files which support things like documents and meta data. This just writes out a list of
conll sentneces to a file in one shot. It is a pretty simple function but I found my self writing it in a lot of scripts so reusing in here
makes sense.