fix: problems in distributed training
Distributed training was passing numpy values in the metrics dict which casued the log output to not be json (should we make this logger
stricter and error on failed JSON conversion? Right now we just fall back to a `__str__` representation which almost looks like json so
it is hard to spot). Distributed is making calls to `.predict`  in the validation and testing steps but it wasn't passing `make_input=False`
so the `.make_input`  function was being called. This caused problems because the data was already processed because we use `tf.data` for
distributed training. Distributed training had problems with going off the end of the iterator so I fixed with with a  for the step count