support no tpu ep, fix cycle/par call mismatch
in ctpu, the TPU resolver can typically resolve the
TPU if an empty EP is given.  Its also possible to query
which TPUs are available with:

`gcloud compute tpus list`

https://cloud.google.com/tpu/docs/troubleshooting

also for now, set cycle_length to the # workers and
autotune the parallel read calls.  the previous setting
was causing errors