improve TB logging in both PyTorch and TF
TensorBoard logging wasnt very informative or
helpful, and our reporting hook was still depending
on tensorboardX which is now integrated into PyTorch.

This change makes the reporting hook use PyTorch's TB
logging if it exists, and if not, fall back to use the
tf library.  This somewhat decouples it from having to
check the backend.

Additionally, adds LR logging in the metrics for PyTorch,
and for TF, where the Keras API means that this doesnt
make much sense, instead log all the LRs in each decay
LRS