fix: batch stability fixes and test for pytorch attention
When trying to tack down the problem of different seq2seq results when
running examples as a batch and as multiple single examples I was seeing
the there are sill occasional errors in both the beam search and the
greedy decoding tests. This suggested there was something before the
decoding that is causing the error.

I tracked this down to the attention implementation of Bahdanau
attention. We weren't actually masking out the attention scores that
were past the end of the example because we were using the out-of-place
version of `masked_fill` but we weren't assigning it to anything.

I feel like I have fixed this (or similar problems in attention) before
so I added batch stability tests so we don't run into problems like this
in the future.

With these changes I ran about 2000 examples (batched into batches of
size 100) through a seq2seq model I trained and there were no
differences between the batched and single version.