fix: set pyt models to eval mode when calling predict (#680)
Currently when we load a pytorch model with a baseline service and call the predict method it often runs in training
mode where drop is turned on which is not what we want at test time. This PR updates the predict method of the tagger,
classifier, and lm to match the predict of the seq2seq model where it is always turned on.

Is there were we want to fix the problem? If someone wanted to do some sort of ensambling of different "draws" of the
model by running multiple times with dropout on they could because we would keep turning it off. The other option is
to set the model to eval mode in the the service, this would me something that would be done based on checking the
backend value