fix: stability of hashing baseline configs
Currently sub dict inside lists were not being sorted. This can causes
differences in the hashes and has now been fixed.

Added the ability to sort specific list values if we control them and
know their order doesn't matter, specifically the `modules` list. It
goes through a set is preprocessing so the order isn't always the same
and was generating multiple hashes.

Added tests for these functions and make sure that sub dictionaries are
getting sorted.