fix: make bert ONNX exportable
The `WordpieceDict1DVectorizer` didn't handle the case where the input
was just a list of tokens that same way the `Dict1DVectorizer`s did
which causes problems because the fake data was always passed in as a
list of tokens.

This updates the `.iterable` method to handle the list of tokens case.
It also updates the method to match the version in
`Wordpiece1DVectorizer`.

This also updates the way labels are read and assigned to tokens in the
`.format_output` method of the taggers. Basically the bert models were
outputing things like the `<PAD>` symbol for the subtokens and we needed
to skip those.

It also bumps the opset we use for ONNX because if you have a lengths
key that goes unused when the ONNX export traces you model then that
input is thrown away in the exported model. Then when you predict with
that key you get an error. Opset 12 will keep that input and remove the error.