chore: Add tests for the new label indices funcs
This PR adds tests to the new vectorizer functionality that tells us
which label indices are valid to use as labels.

If we want we could refactor this in the future to actually use the
tokenizers to tokenize random text and track how they are broken instead
of generating fake data by manually breaking like I am doing here.