make WeightTiedDense lazy via a weight property (#709)
* make WeightTiedDense lazy via a weight property

this change makes weight tying module use the
tied module's underlying weight directly instead of
copying a pointer parameter.  Previously we were using a
pointer, which causes problems when the module is
rehydrated from disk after first initialization.

This was found because the number of parameters in the
model was increasing after embeddings were reloaded,
which were tied to an LM head.  One other nice property of
deferring to the "held" module is that we dont need properties
instructing us reload the LM head, because this will happen
automatically for models with a tied head, simplifying TLM
(and TiedSeq2Seq) reloads from checkpoints

The only concern with this approach is that there is now
a function powering the `weight` which may cause a bit of latency,
though compared to the matrix multiply this is probably negligable.
Prior to this change however, there was (at a minimum) some memory
that shouldve been freeable that was still being tracked in PyTorch,
and at worst, may have been affecting the NN fine-tuning

* try and fix TF one as well