fix: conlleval updates
There was a type in the `_get_entities` function name (forgot the `i`).
It was missing in both the definition and the call so it worked but it
was bugging me.

This updates the default delimiter too. It was set to be an actual span
`" "` but this makes seemed overly strict, by defaulting it to `None` we
can automatically handle both tab separated and space separated columns.