update to support paired classification
oops, remove the break, vec with no [CLS]

remove isinstance

update TF side to support paired models

oops

try and support TPUs

hotshot commit

refactor all to use create_distribute_strategy

tell torch we dont need to use all that memory

make pytorch TT embeddings have no pad idx

make TF work with token_type

add support for skipping the vocab counting

this was added in master and so this is rebased

fix issue with label file

fix reader bug and add cased nocls vecs

add a label list option+refactor to make available

we were returning only the file portions of the metadata
from the DataDownloader, which was weird and quirky.
what we really want to do is overlay and update file paths
but return the whole md block to the task.

the task should then inject the full block into the reader so
the reader can access all the fields.  then its easy to add
new fields in that get propagated, like `label_list`.  Refactored
the readers to try to read from `label_file` but also `label_list`.
this makes it possible to add metadata directly into the index without
needing to download a whole different tarball.

fix offsets when pre-reading labels for tagging

im not really sure we should even continue to support
the use case of reading in labels for taggers.  it seems
very error prone and not expensive to read on the fly

+non-prefix/suffix WP tokenizers refactor Task

added some non-CLS,non-SEP vectorizers, refactor the
Task to provide a base function for Backend setup.
The part that is not-being repeated seems to actually
belong in the caller `_setup_task()`

quick hack for TPUs

more hacking

monkeys compose shakespeare

hack FS

try to allow it to live on GCP

remove model save for now, redundant with checkpoint

add SNLI config