define interface for predefined to optimize reads (#726)
when we have a predefined vocab, there is no point in
tabulating the vocabulary.  this is especially a problem
for large classification, seq2seq and LM datasets where we
end up doing a pointless compute.  this change defines an
interface that is implemented by vectorizers with predefined
vocabs to facilitate skipping this tabulation on reads