add optional training preproc buckets to TF example
When training, we sometimes would like to pass
in bucketed data by size.  To keep the training
curve smooth, this should either scale down the
LR for smaller length buckets or it should scale
up the batch size.

Since we preprocess all the data in this sample, this
comes down to maintaining some knowledge of the width
of each context window and the relationship to the
batch size and LR for the (presumably) largest context
length

This change is backwards compatible but supports
an alternate layout for the training data with
sub-directories named for the context length of
each bucket, and an updated `md.yml` that provides
the sample for each bucket