added support for subword-nmt package (#746)
previously, we supported Lample's fastBPE, which
is a fast C++ implementation of Rico Sennrich's BPE
software (subword-nmt).  The subword-nmt package
supports some train time features like BPE-dropout,
and it supports glossaries during BPE application.

this adds support for subword NMT as an alternate
provider.  This support is also compatible with fastBPE-
trained models, which we achieve by inheriting the base
package's BPE class and modifying the codes reader.

The original BPE can be accessed by setting `use_fast_bpe`
to `False`, and new words can be added by passing in
`glossaries` as either a list or a file