1. use common dist strat fn 2. monkey patch run
TF 2.1 includes the experimental_run_v2 function from
the strategy class, but by 2.4 this is gone entirely.
the easiest way to get around this seems to be to monkey
patch tf < 2.2 to create a run which is the former, and
to update each instance of run to the new method.

To make this easier, centralize this inside the
common layers create_distribute_strategy, which necessitates
updating the sample API programs to use this.  There was a
weird arg mismatch between the library function and the samples
but since the samples are used prominently, I switched the lib arg
order to match