demo how to fix service for no text
we normally assume there are text features in the input
its possible to not do this, but in that case 2 updates are
needed.

the first is that, because we assume there is text, we do
some things in the service to backout things like BPE
tokenization/vectorization which yields a bunch of PAD
tokens that we ignore in the service output.

the second is that the tag_text.py program expects that
there is going to be a word-label combo when it prints the
output, so make that optional here.

I kind of conflated these 2 into a single option for the
tag_text.py demo.  If you pass `labels_only True` it will
assume that we dont want `valid_labels_only` in the output,
and it will also assume we dont want to print text.  This is
probably all right for now, but might be worth revisiting as
they are technically separate