one path to reload for DE and S2S
this cleans up the reload of checkpoints so they
are unified between seq2seq and dual-encoder.

it makes it possible to pass in a directory, in which case
we find the latest checkpoint

for S2S reloading, it only updates the global step/epoch num
when it reloads either a PTH or a NPZ file for transformer
seq2seq models.  the backoff behavior remains, where we try and
reload the encoder from a transformer LM NPZ but this should be
treated as step 0, not whatever step was given in the checkpoint

additionally, for DEs, they only preserve step when reloading
a PTH, for NPZ reloads it use the transformer LM NPZ and loads
the .transformer and .embeddings only, so again, it should be
treated as step 0