fixes learned positional with bias inputs on save
BERT has 3 types of embeddings that are added together,
the usual LUT word (piece), the learned positional and the
token type.  we support LP embeddings, and we support using
token type as a second embedding which is added in the reduction
`sum-layer-norm` to execute BERT.  However, the token type embeddings
are not really useful in cases where the input is not structured
as cross-attention (a single vector for a pair of data).  In other
cases, the token-type embeddings just get set to the 0 value, which
is constant when a BERT checkpoint is reloaded.

LP w/ bias is an optimization where we load the 0 vector for TT and
we always just add that term to the LP embeddings.  This works great
except we werent saving these back out.

this change is to save those LP w/ bias embeddings in the TLM NPZ format
where they can be read back in.  Important note, as we are (presumably)
fine-tuning BERT in these cases in a way that is not updating the token
type, in fact, as we have thrown away the input #2, the 2nd token type
here recieves a zero value.  This would be a problem for reuse inside
a full token type embedding.  IOW, this change allows us to save to the
format, but it does not fix reloading them once they are converted into
a singe bias term, as that information is lost