Add tri-stage LRS (#785)
* Add tri-stage LRS

implemented by an optional `plateau_steps` in
the composite LRS which defaults to 0.  The warmup
LRS is guaranteed to max out its scale factor at the
number of warmup steps, so we can modify the composite
to continue to call the warmup LRS until both `warmup_steps`
and `plateau_steps` are exceeded.  Then we adjust the final
decay scheduler accordingly.

* bump version