this fixes an issue using known_vocab Counter
the code was changed in PR #796 to make it easier
to programmatically inject a vocab in using `known_vocab`
but this is a magic keyword arg that holds counters mostly
during `mead-train` which causes a catastrophic issue during
training where the indices are changed to counts.  The
previous behavior was to back off to a global var loaded by
the vectorizers during `mead-train` which is obnoxious as well
but is the previous behavior.

The fix here should support both solutions by testing if the vocab
is either null or a Counter so it can then use the global solution