Fix Offsets on reload by __setstate__ overload
Unpickling doesnt call the constructors for vectorizers
that mutate the global Offsets.  These Offsets have to be
modified for functions like `pads()` to work as expected
at inference time, since some models (namely RoBERTa)
dont use the same offsets for PAD, GO, EOS etc.

All the other values are persisted in the class so they
shouldnt be a problem, but the Offsets are global so
this change is lost