safety check to make sure records are valid
normally this isnt a problem, but if the
data gets truncated, it may be necessary to
throw the sample away in rare cases.

here we add an option to test if the final
record matches our expectation, and if not,
we will throw it out.

currently this behavior only affects the paired
preproc program although it might be also
desirable in other cases