minor mod to support loading all features
there is a potential issue with this approach, having
to do with lengths of the models.  If there are multiple
vectors with different lengths, this could be an issue,
or if there are any scalars.  A simple way to handle this
should be to use the "primary feature" but its not clear
if thats being preserved here.  for now its coded like it is