Feature/add offsets to spm (#911)
* changes the base SPM vectorizer to add offsets

in this branch for now, we default this to true and
GPT2 style tokenization to false.  Need to come up
with a way to pass this info through which doesnt
affect other impls

* smooth out edges for default of vectorizer

change the SPM vectorizer to have a magic
{{OFFSETS}} value that can be passed for
`--extra_tokens`.  Add logic to rewire the
offset values if that should be used