refactor transformer layer -> easier understanding
Right now we try to handle all paths through the
same path.  This is error prone and difficult to reading.
Compounding this, our pre-LN models to date are non-std,
which means its even harder to get whats going on.

This refactoring pulls up an ABC for both, and provides
separate implementations for all 3 that are supported
preLN, postLN, TPT style